(Commit 1 by Oriol)
Created Pokedex POJO
Created Pokemon POJO
Created database generator.

The app downloads the information from the pokeapi JSON files and saves it into a SQLite database to show it in a ListView.

Also shows details and image if you click on one of the Pokemons.

Added the search option to search by name.